+++ packages/react-storage/src/components/StorageBrowser/actions/index.ts
@@ -1,26 +1,10 @@
 export {
-  ActionConfigs,
-  ActionConfigsProvider,
-  ActionConfigsProviderProps,
-  ComponentName,
-  defaultActionConfigs,
-  DefaultActionConfigs,
-  defaultActionViewConfigs,
-  DefaultActionKey,
-  isDefaultActionViewType,
-  SelectionType,
-  TaskActionConfig,
-  useActionConfig,
-} from './configs';
-
-export {
   ActionInputConfig,
   copyHandler,
   CopyHandler,
   CopyHandlerData,
   CopyHandlerInput,
   CopyHandlerOutput,
-  createFileDataItemFromLocation,
   createFileDataItem,
   createFolderHandler,
   CreateFolderHandler,
@@ -65,6 +49,7 @@ export {
   ListLocationsHandlerOutput,
   LocationData,
   LocationItemData,
+  LocationItemType,
   LocationPermissions,
   LocationType,
   TaskData,
@@ -80,6 +65,18 @@ export {
   UploadHandlerOutput,
 } from './handlers';
 
-export { ActionState } from './types';
-
-export { useListLocations, UseListLocationsState } from './useAction';
+export {
+  ExtendedActionConfigs,
+  ActionViewConfig,
+  ActionViewConfigs,
+  ActionConfigsProvider,
+  ActionConfigsProviderProps,
+  ActionHandler,
+  CustomActionConfigs,
+  defaultActionConfigs,
+  DefaultActionConfigs,
+  defaultActionViewConfigs,
+  getActionConfigs,
+  isDefaultActionViewType,
+  useActionConfigs,
+} from './configs';