+++ packages/react-storage/src/components/StorageBrowser/controls/__tests__/NavigationControl.spec.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import { useResolvedComposable } from '../hooks/useResolvedComposable';
-import { useNavigation } from '../hooks/useNavigation';
 import { NavigationControl } from '../NavigationControl';
+import { useNavigation } from '../hooks/useNavigation';
+import { useResolvedComposable } from '../hooks/useResolvedComposable';
 
 jest.mock('../hooks/useNavigation');
 jest.mock('../hooks/useResolvedComposable');
+jest.mock('../../composables/Navigation', () => ({
+  Navigation: () => <div data-testid="navigation" />,
+}));
 
 describe('NavigationControl', () => {
-  // assert mocks
   const mockUseNavigation = jest.mocked(useNavigation);
   const mockUseResolvedComposable = jest.mocked(useResolvedComposable);
-  // create mocks
-  const mockOnNavigate = jest.fn();
 
   beforeAll(() => {
     mockUseResolvedComposable.mockImplementation(
@@ -21,24 +21,14 @@ describe('NavigationControl', () => {
   });
 
   afterEach(() => {
-    mockUseNavigation.mockReset();
-    mockUseResolvedComposable.mockClear();
-    mockOnNavigate.mockClear();
+    mockUseNavigation.mockClear();
   });
 
   it('renders', () => {
-    mockUseNavigation.mockReturnValue({
-      items: [
-        { name: 'Item 1', onNavigate: mockOnNavigate },
-        { name: 'Item 2', onNavigate: mockOnNavigate, isCurrent: true },
-      ],
-    });
-
     render(<NavigationControl />);
 
-    const [item1, item2] = screen.getAllByRole('listitem');
+    const navigation = screen.getByTestId('navigation');
 
-    expect(item1).toHaveTextContent('Item 1');
-    expect(item2).toHaveTextContent('Item 2');
+    expect(navigation).toBeInTheDocument();
   });
 });