+++ packages/react-native-auth/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @aws-amplify/ui-react-native-auth
 
+## 0.0.37
+
+### Patch Changes
+
+- Updated dependencies [[`9ab56f499f4596062fe4614a016efc3dfc88ef03`](https://github.com/aws-amplify/amplify-ui/commit/9ab56f499f4596062fe4614a016efc3dfc88ef03)]:
+  - @aws-amplify/ui@6.7.1
+  - @aws-amplify/ui-react-core-auth@0.0.34
+  - @aws-amplify/ui-react-native@2.2.19
+
+## 0.0.36
+
+### Patch Changes
+
+- Updated dependencies [[`70971f310`](https://github.com/aws-amplify/amplify-ui/commit/70971f310f69717657849f29d0a9e1d993b08d9a)]:
+  - @aws-amplify/ui@6.7.0
+  - @aws-amplify/ui-react-core-auth@0.0.33
+  - @aws-amplify/ui-react-native@2.2.18
+
 ## 0.0.35
 
 ### Patch Changes